+++ src/main/kotlin/site/siredvin/turtlematic/computercraft/ComputerCraftProxy.kt
@@ -24,6 +24,7 @@ import site.siredvin.turtlematic.computercraft.peripheral.misc.*
 import site.siredvin.turtlematic.computercraft.turtle.BlockTurtleUpgrade
 import site.siredvin.turtlematic.computercraft.turtle.ClockwiseAnimatedTurtleUpgrade
 import site.siredvin.turtlematic.computercraft.turtle.StarboundTurtleUpgrade
+import site.siredvin.turtlematic.computercraft.turtle.TickerFunctions
 
 object ComputerCraftProxy {
 
@@ -41,11 +42,23 @@ object ComputerCraftProxy {
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.AUTOMATA_CORE) { turtle, side, tier -> AutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
-        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.NETHERITE_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.ticker(
+            Items.NETHERITE_HUSBANDRY_AUTOMATA_CORE,
+            { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) },
+            TickerFunctions::netheriteHusbandryTick
+        ))
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.NETHERITE_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
-        TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(StarboundTurtleUpgrade.ticker(
+            Items.STARBOUND_HUSBANDRY_AUTOMATA_CORE,
+            { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) },
+            TickerFunctions::starboundHusbandryTick
+        ))
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.STARBOUND_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
-        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_HUSBANDRY_AUTOMATA_CORE) { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) })
+        TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.ticker(
+            Items.CREATIVE_HUSBANDRY_AUTOMATA_CORE,
+            { turtle, side, tier -> HusbandryAutomataCorePeripheral(turtle, side, tier) },
+            TickerFunctions::creativeHusbandryTick
+        ))
         TURTLE_UPGRADES.add(ClockwiseAnimatedTurtleUpgrade.dynamic(Items.CREATIVE_END_AUTOMATA_CORE) { turtle, side, tier -> EndAutomataCorePeripheral(turtle, side, tier) })
         TURTLE_UPGRADES.add(StarboundTurtleUpgrade.dynamic(Items.ENORMOUS_AUTOMATA_CORE) { turtle, side, tier -> EnormousAutomataCorePeripheral(turtle, side, tier) })